Reorganized testing file structure approach
