+++ jaseci_core/jaseci/api/tests/fixtures/walker_yield.jac
@@ -1,0 +1,10 @@
+walker init {
+    has a=6;
+
+    if(a < 10) {
+        a+=1;
+        report a;
+        yield;
+    }
+}
+
