+++ jaseci_core/jaseci/api/tests/test_walker_api.py
@@ -1,0 +1,25 @@
+from .test_api_core import core_test
+
+
+class walker_api_test(core_test):
+    """Unit tests for Jac Walker APIs"""
+
+    def test_another_thing(self):
+        api = ["graph_create", {}]
+        r = self.call(self.mast, api)
+        before = len(self.mast._h.mem)
+        jid = r["jid"]
+        api = ["graph_node_set", {"nd": jid, "ctx": {}}]
+        r = self.call(self.mast, api)
+        api = ["graph_node_set", {"nd": jid, "ctx": {}}]
+        r = self.call(self.mast, api)
+        api = ["graph_node_set", {"nd": jid, "ctx": {}}]
+        r = self.call(self.mast, api)
+        api = ["graph_node_set", {"nd": jid, "ctx": {}}]
+        r = self.call(self.mast, api)
+        api = ["graph_node_set", {"nd": jid, "ctx": {}}]
+        r = self.call(self.mast, api)
+        api = ["graph_node_set", {"nd": jid, "ctx": {}}]
+        r = self.call(self.mast, api)
+        after = len(self.mast._h.mem)
+        self.assertEqual(before, after)
